+++ src/encoded/types/higlass_view_config.py
@@ -1,7 +1,8 @@
 from copy import deepcopy
 from snovault import collection
 from encoded_core.types.higlass_view_config import HiglassViewConfig as CoreHiglassViewConfig
-from .base import SMAHTItem, mixin_smaht_permission_types
+from .base import Item as SMAHTItem
+from .base import mixin_smaht_permission_types
 
 
 ENCODED_CORE_HIGLASS_VIEW_CONFIG_SCHEMA = deepcopy(CoreHiglassViewConfig.schema)